refactor(examples): extract common helpers module
